+++ src/filament_manager/__init__.py
@@ -1,3 +1,3 @@
 """Filament Manager application package."""
 
-__version__ = "0.6.2"
+__version__ = "0.6.3"